+++ src/js/features/exercises/components/creator.jsx
@@ -1,29 +1,36 @@
 import React, { Fragment, useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { createNewExercisePost } from '../state/actions';
+import { getExerciseList, putExercise } from '../state/actions';
 
 const propTypes = {
-  createNewExercisePost: PropTypes.func,
-  posting: PropTypes.bool,
-  successful: PropTypes.bool
+  exercises: PropTypes.array,
+  getExerciseList: PropTypes.func,
+  isLoading: PropTypes.bool,
+  isProcessing: PropTypes.bool,
+  putExercise: PropTypes.func,
+  successfulPut: PropTypes.bool
 };
-const exerciseCreator = props => {
-  const submitButtonId = 'exerciseCreatorSubmitButton';
+const exerciseCreator = ({ exercises, getExerciseList, isLoading, isProcessing, putExercise, successfulPut }) => {
   const [description, setDescription] = useState('');
+  const [isSubmitDisabled, setIsSubmitDisabled] = useState(false);
   const [name, setName] = useState('');
   const [typeId, setTypeId] = useState('');
 
   useEffect(() => {
-    if (props.posting) {
-      document.getElementById(submitButtonId).disabled = true;
-    } else {
-      document.getElementById(submitButtonId).disabled = false;
-      if (props.successful) {
-        resetFormValues();
-      }
+    resetFormValues();
+  }, []);
+
+  useEffect(() => {
+    setIsSubmitDisabled(isLoading || isProcessing);
+  }, [isLoading, isProcessing]);
+
+  useEffect(() => {
+    if (successfulPut) {
+      getExerciseList();
+      resetFormValues();
     }
-  }, [props.posting]);
+  }, [successfulPut]);
 
   const resetFormValues = () => {
     setDescription('');
@@ -31,39 +38,40 @@ const exerciseCreator = props => {
     setTypeId(1);
   };
 
-  const handleSubmit = event => {
-    event.preventDefault();
-    const payload = { description, name, typeId };
-    props.createNewExercisePost(payload);
+  const handleSubmit = () => {
+    const id = 'ex-id-' + new Date().getTime();
+    const item = { description, id, name, typeId };
+    const payload = { exercises: [item, ...exercises] };
+    putExercise(payload);
   };
 
   return (
     <Fragment>
       <p><strong>Create a new Exercise</strong></p>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label forhtml="name">Name</label><br/>
-          <input name="name" onChange={event => setName(event.target.value) } type="text" value={name} />
-        </div>
-        <div>
-          <label forhtml="description">Description</label><br/>
-          <textarea name="description" onChange={event => setDescription(event.target.value)} value={description}>
-          </textarea>
-        </div>
-        <div>
-          <label forhtml="typeId">Type</label><br/>
-          <input name="typeId" onChange={event => setTypeId(Number(event.target.value)) } type="number" value={typeId} />
-        </div>
-        <div>
-          <input id={submitButtonId} type="submit" value="Submit" />
-        </div>
-      </form>
+      <div>
+        <label forhtml="name">Name</label><br/>
+        <input name="name" onChange={event => setName(event.target.value) } type="text" value={name} />
+      </div>
+      <div>
+        <label forhtml="description">Description</label><br/>
+        <textarea name="description" onChange={event => setDescription(event.target.value)} value={description}>
+        </textarea>
+      </div>
+      <div>
+        <label forhtml="typeId">Type</label><br/>
+        <input name="typeId" onChange={event => setTypeId(Number(event.target.value)) } type="number" value={typeId} />
+      </div>
+      <div>
+        <button disabled={isSubmitDisabled} role="button" onClick={handleSubmit}>Submit</button>
+      </div>
     </Fragment>);
 };
 
 const mapStateToProps = state => ({
-  posting: !!state.exercises.posting,
-  successful: !!state.exercises.successfulPost
+  exercises: state.exercises.exerciseGetPayload,
+  isLoading: state.exercises.isLoadingExercises,
+  isProcessing: state.exercises.isProcessingExercise,
+  successfulPut: !!state.exercises.exercisePostPayload
 });
 exerciseCreator.propTypes = propTypes;
-export default connect(mapStateToProps, { createNewExercisePost })(exerciseCreator);
+export default connect(mapStateToProps, { getExerciseList, putExercise })(exerciseCreator);