+++ src/js/features/exercises/index.jsx
@@ -1,14 +1,21 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import { getExerciseList } from './state/actions';
 import Card from '../../components/bulmaCard';
 import Layout from '../../components/layout';
 import ExerciseCreator from './components/creator';
 import ExerciseDropdown from './components/dropdown';
 
-const exercisePage = () => {
-  const exerciseCreatorChild = (<ExerciseCreator />);
-  const exerciseCreatorTitle = 'Exercise Creator';
-  const exerciseDropdownChild = (<ExerciseDropdown />);
-  const exerciseDropdownTitle = 'Exercise Dropdown Sample';
+const propTypes = {
+  exercises: PropTypes.array,
+  getExerciseList: PropTypes.func
+};
+const exercisePage = ({ exercises, getExerciseList }) => {
+  useEffect(() => {
+    getExerciseList();
+  }, []);
+
   return (
     <Layout title='Exercises'>
       <div className="card">
@@ -17,14 +24,15 @@ const exercisePage = () => {
         </header>
         <div className="card-content columns is-tablet">
           <div className="content column is-one-half">
-            <Card child={exerciseCreatorChild} title={exerciseCreatorTitle} />
+            <Card child={<ExerciseCreator />} title="Exercise Creator" />
           </div>
           <div className="content column is-one-half">
-            <Card child={exerciseDropdownChild} title={exerciseDropdownTitle} />
+            <Card child={<ExerciseDropdown exercises={exercises} />} title="Exercise Dropdown Sample" />
           </div>
         </div>
       </div>
     </Layout>);
 };
-
-export default exercisePage;
+exercisePage.propTypes = propTypes;
+const mapStateToProps = state => ({ exercises: state.exercises.exerciseGetPayload });
+export default connect(mapStateToProps, { getExerciseList })(exercisePage);