+++ src/js/features/nutrients/caffeine/components/creator.jsx
@@ -1,31 +1,39 @@
 import React, { Fragment, useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { createNewCaffieneConsumption } from '../state/actions';
+import { getCaffeineList, putCaffeine } from '../state/actions';
 import '../../../../../css/creator.css';
 
 const propTypes = {
-  createNewCaffieneConsumption: PropTypes.func,
-  posting: PropTypes.bool,
-  successful: PropTypes.bool
+  caffeine: PropTypes.array,
+  getCaffeineList: PropTypes.func,
+  isLoadingCaffeine: PropTypes.bool,
+  isProcessingCaffeine: PropTypes.bool,
+  isSuccessfulPut: PropTypes.bool,
+  putExercise: PropTypes.func
 };
-const creator = props => {
-  const submitButtonId = 'submitNutrientConsumption';
+const creator = ({ caffeine, getCaffeineList, isLoadingCaffeine, isProcessingCaffeine, isSuccessfulPut, putCaffeine }) => {
   const [amount, setAmount] = useState(0);
   const [amountType, setAmountType] = useState('');
   const [comment, setComment] = useState('');
+  const [isSubmitDisabled, setIsSubmitDisabled] = useState(false);
   const [userName, setUserName] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
-    if (props.posting) {
-      document.getElementById(submitButtonId).disabled = true;
-    } else {
-      document.getElementById(submitButtonId).disabled = false;
-      if (props.successful) {
-        resetFormValues();
-      }
+    setIsLoading(isLoadingCaffeine || isProcessingCaffeine);
+  }, [isLoadingCaffeine, isProcessingCaffeine]);
+
+  useEffect(() => {
+    setIsSubmitDisabled(isLoading);
+  }, [isLoading]);
+
+  useEffect(() => {
+    if (isSuccessfulPut) {
+      resetFormValues();
+      getCaffeineList();
     }
-  }, [props.posting]);
+  }, [isSuccessfulPut]);
 
   const resetFormValues = () => {
     setAmount(0);
@@ -34,8 +42,7 @@ const creator = props => {
     setUserName('');
   };
 
-  const handleSubmit = event => {
-    event.preventDefault();
+  const handleSubmit = () => {
     const payload = {
       amount: Number(amount),
       amountType: amountType,
@@ -43,56 +50,56 @@ const creator = props => {
       intakeTime: new Date().toJSON(),
       userName: userName
     };
-    props.createNewCaffieneConsumption(payload);
+    putCaffeine({ caffeine: [payload, ...caffeine] });
   };
 
   return (
     <Fragment>
-      <form onSubmit={handleSubmit}>
-        <div className="field is-horizontal">
-          <label className="label">Amount</label>
-          <input className="input input-amount"
-            name="amount"
-            onChange={ e => setAmount(e.target.value) }
-            required
-            type="number"
-            value={amount} />
-          <label className="label">Type</label>
-          <select onChange={ event => setAmountType(event.target.value)} value={amountType}>
-            <option default value="">--</option>
-            <option value="mg">Milligram</option>
-            <option value="kg">Kilogram</option>
-            <option value="pound">Pound</option>
-            <option value="ounce">Ounce</option>
-          </select>
-        </div>
-        <div className="field is-horizontal">
-          <label className="label">User</label>
-          <input className="input"
-            name="userName"
-            onChange={ event => setUserName(event.target.value) }
-            required
-            type="text"
-            value={userName} />
-        </div>
-        <div className="field">
-          <label className="label">Comment</label>
-          <textarea className="textarea"
-            name="comment"
-            onChange={ event => setComment(event.target.value) }
-            type="textarea"
-            value={comment} />
-        </div>
-        <div>
-          <input id={submitButtonId} type="submit" value="Submit" />
-        </div>
-      </form>
+      <div className="field is-horizontal">
+        <label className="label">Amount</label>
+        <input className="input input-amount"
+          name="amount"
+          onChange={ e => setAmount(e.target.value) }
+          required
+          type="number"
+          value={amount} />
+        <label className="label">Type</label>
+        <select onChange={ event => setAmountType(event.target.value)} value={amountType}>
+          <option default value="">--</option>
+          <option value="mg">Milligram</option>
+          <option value="kg">Kilogram</option>
+          <option value="pound">Pound</option>
+          <option value="ounce">Ounce</option>
+        </select>
+      </div>
+      <div className="field is-horizontal">
+        <label className="label">User</label>
+        <input className="input"
+          name="userName"
+          onChange={ event => setUserName(event.target.value) }
+          required
+          type="text"
+          value={userName} />
+      </div>
+      <div className="field">
+        <label className="label">Comment</label>
+        <textarea className="textarea"
+          name="comment"
+          onChange={ event => setComment(event.target.value) }
+          type="textarea"
+          value={comment} />
+      </div>
+      <div>
+        <button disabled={isSubmitDisabled} role="button" onClick={handleSubmit}>Submit</button>
+      </div>
     </Fragment>);
 };
 
 const mapStateToProps = state => ({
-  posting: !!state.exercises.posting,
-  successful: !!state.exercises.successfulPost
+  caffeine: state.caffeineIntakes.caffeineGetPayload,
+  isLoadingCaffeine: state.caffeineIntakes.isLoadingCaffeine,
+  isProcessingCaffeine: state.caffeineIntakes.isProcessingCaffeine,
+  isSuccessfulPut: !!state.caffeineIntakes.caffeinePostPayload
 });
 creator.propTypes = propTypes;
-export default connect(mapStateToProps, { createNewCaffieneConsumption })(creator);
+export default connect(mapStateToProps, { getCaffeineList, putCaffeine })(creator);