+++ src/js/features/nutrients/caffeine/state/initialState.js
@@ -1,10 +1,8 @@
 export default {
-  currentPage: 0,
-  didInvalidate: true,
-  intakes: [],
-  isFetching: false,
-  links: { first: '', last: '', next: '', prev: '', self: '' },
-  totalPages: 0,
-  totalRecords: 0,
-  lastUpdated: null
+  isLoadingCaffeine: false,
+  isProcessingCaffeine: false,
+  caffeineGetPayload: null,
+  caffeineGetError: null,
+  caffeinePostPayload: null,
+  caffeinePostError: null
 };