+++ src/js/features/sessions/components/table.jsx
@@ -1,52 +1,35 @@
-import React, { Fragment, useEffect } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import ModalButton from './modalButton';
-import TablePagination from '../../../components/tablePagination';
-import { fetchSessionsIfNeeded } from '../state/actions';
+import { getCardioSessionList } from '../state/actions';
 import { splitTextKeyToArray } from '../../../lib/splits';
 import * as _config from '../../../../assets/config.json';
 import '../../../../css/table.css';
 
-const apiUrl = _config.apis.azure + 'CardioMachineExercises?pageNumber=1&pageSize=10';
 const propTypes = {
-  currentPage: PropTypes.number,
-  didInvalidate: PropTypes.bool,
-  fetchSessionsIfNeeded: PropTypes.func,
-  isFetching: PropTypes.bool,
-  links: PropTypes.object,
-  sessions: PropTypes.array,
-  totalPages: PropTypes.number
+  getCardioSessionList: PropTypes.func,
+  isLoadingSession: PropTypes.bool,
+  isProcessingSession: PropTypes.bool,
+  sessions: PropTypes.array
 };
+const table = ({ getCardioSessionList, isLoadingSession, isProcessingSession, sessions }) => {
+  const cardioSessionGetPath = 'json/object/cardio';
+  const downloadText = 'Download Session in JSON file.';
 
-const table = props => {
+  const [isLoading, setIsLoading] = useState(false);
+
   useEffect(() => {
-    props.fetchSessionsIfNeeded(apiUrl);
+    getCardioSessionList();
   }, []);
 
   useEffect(() => {
-    if (props.didInvalidate) {
-      props.fetchSessionsIfNeeded(apiUrl);
-    }
-  }, [props.didInvalidate]);
-
-  const renderDownloadLink = () => {
-    const url = _config.apis.azure + 'CardioMachineExercises?csv=csv';
-    return <a className="card-footer-item" href={url}>Download Intakes as CSV file.</a>;
-  };
-
-  const renderPagination = () => {
-    return (props.isFetching)
-      ? <div className='circular-loader'><CircularProgress /></div>
-      : (<TablePagination currentPage={props.currentPage}
-        links={props.links}
-        totalPages={props.totalPages}
-      />);
-  };
+    setIsLoading(isLoadingSession || isProcessingSession);
+  }, [isLoadingSession, isProcessingSession]);
 
   const renderSessionsData = () => {
-    return Array.isArray(props.sessions) && props.sessions.map((element, index) => {
+    return Array.isArray(sessions) && sessions.map((element, index) => {
       const date = splitTextKeyToArray(element, 'startTime', 'T')[0];
       return (
         <tr key={index}>
@@ -59,7 +42,7 @@ const table = props => {
   };
 
   const renderTableRows = () => {
-    return (!props.isFetching && !!props.sessions) && (<tbody>{renderSessionsData()}</tbody>);
+    return (!isLoading && !!sessions) && (<tbody>{renderSessionsData()}</tbody>);
   };
 
   const renderTableHeader = () => {
@@ -76,18 +59,15 @@ const table = props => {
           {renderTableRows()}
         </table>
       </div>
-      {renderPagination()}
-      {renderDownloadLink()}
+      <a className="card-footer-item" href={_config.apis.heroku + cardioSessionGetPath} target="_">{downloadText}</a>
+      {isLoading && <div className='circular-loader'><CircularProgress /></div>}
     </Fragment>);
 };
 
 table.propTypes = propTypes;
 const mapStateToProps = state => ({
-  currentPage: state.cardioMachineSessions.currentPage,
-  didInvalidate: state.cardioMachineSessions.didInvalidate,
-  isFetching: state.cardioMachineSessions.isFetching,
-  links: state.cardioMachineSessions.links,
-  sessions: state.cardioMachineSessions.sessions,
-  totalPages: state.cardioMachineSessions.totalPages
+  isLoadingSession: state.cardioMachineSessions.isLoadingCardioSessions,
+  isProcessingSession: state.cardioMachineSessions.isProcessingCardioSession,
+  sessions: state.cardioMachineSessions.cardioSessionGetPayload
 });
-export default connect(mapStateToProps, { fetchSessionsIfNeeded })(table);
+export default connect(mapStateToProps, { getCardioSessionList })(table);