+++ src/js/features/sessions/index.jsx
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types';
 import Page from '../../components/page';
 import SessionCreator from './components/creator';
 import SessionsTableDisplay from './components/table';
-import { fetchSessionsIfNeeded } from './state/actions';
+import { getCardioSessionList } from './state/actions';
 
-const propTypes = { fetchSessionsIfNeeded: PropTypes.func.isRequired };
-const sessions = props => {
+const propTypes = { getCardioSessionList: PropTypes.func };
+const sessions = ({ getCardioSessionList }) => {
   const tabNames = ['Information', 'Record', 'History', 'Data Visualization'];
   const tabPanels = [
     <p key={1}>Information on fitness sessions</p>,
@@ -17,11 +17,11 @@ const sessions = props => {
   ];
 
   useEffect(() => {
-    props.fetchSessionsIfNeeded();
+    getCardioSessionList();
   }, []);
 
   return <Page title="Sessions" tabNames={tabNames} tabPanels={tabPanels} />;
 };
 
 sessions.propTypes = propTypes;
-export default connect(null, { fetchSessionsIfNeeded })(sessions);
+export default connect(null, { getCardioSessionList })(sessions);