+++ src/js/features/sessions/state/creator/reducer.js
@@ -5,16 +5,7 @@ const reducer = (state = initialState, action) => {
   let newState;
   switch (action.type) {
     case _types.UPDATE_CARDIO_MACHINE_SESSION_POST_REQUEST:
-      newState = { ...state, posting: false, form: { ...state.form, ...action.update } };
-      break;
-    case _types.RECIEVE_CARDIO_MACHINE_SESSION_POST_RESPONSE:
-      newState = { ...state, posting: false };
-      break;
-    case _types.RECIEVE_SUCCESSFUL_CARDIO_MACHINE_SESSION_POST_RESPONSE:
-      newState = { ...initialState, posting: false, successfulPost: true };
-      break;
-    case _types.SEND_CARDIO_MACHINE_SESSION_POST_REQUEST:
-      newState = { ...state, posting: true, successfulPost: false };
+      newState = { ...state, form: { ...state.form, ...action.update } };
       break;
     default:
       newState = { ...state };