+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] -2021-03-20
+## Changed
+- Caffeine, Cardio Session, and Exercises list and create API using S3 storage API
+
+## Removed
+- Pagination on Caffeine and Cardio Session tables
+- Azure API URL from application
+
 ## [0.10.1] -2021-03-07
 ### Added
 - Load indicator on Routine sidebar card